rename api key file and update api key auth implementation
